refactor all participant-hooks into separate files
