+++ packages/react/src/hooks/index.ts
@@ -1,9 +1,21 @@
-export * from './participant-hooks';
-export { useToken, UseTokenOptions } from './useToken';
 export { useDataChannel } from './useDataChannel';
-export { useTracks } from './useTracks';
 export { useGridLayout } from './useGridLayout';
-export { useVisualStableUpdate } from './useVisualStableUpdate';
-export { usePagination } from './usePagination';
+export { useIsMuted, UseIsMutedOptions } from './useIsMuted';
+export { useIsSpeaking } from './useIsSpeaking';
+export { useLocalParticipant } from './useLocalParticipant';
+export { useLocalParticipantPermissions } from './useLocalParticipantPermissions';
 export { useMediaTrack } from './useMediaTrack';
 export { useMediaTrackByName } from './useMediaTrackByName';
+export { usePagination } from './usePagination';
+export {
+  useParticipantPermissions,
+  UseParticipantPermissionsOptions,
+} from './useParticipantPermissions';
+export { useParticipants, UseParticipantsOptions } from './useParticipants';
+export { useRemoteParticipant, UseRemoteParticipantOptions } from './useRemoteParticipant';
+export { useRemoteParticipants, UseRemoteParticipantsOptions } from './useRemoteParticipants';
+export { useSortedParticipants } from './useSortedParticipants';
+export { useSpeakingParticipants } from './useSpeakingParticipants';
+export { useToken, UseTokenOptions } from './useToken';
+export { useTracks } from './useTracks';
+export { useVisualStableUpdate } from './useVisualStableUpdate';
