+++ packages/react/src/hooks/useIsMuted.ts
@@ -1,0 +1,20 @@
+import { mutedObserver } from '@livekit/components-core';
+import { Participant, Track } from 'livekit-client';
+import * as React from 'react';
+import { useEnsureParticipant } from '../context';
+
+export interface UseIsMutedOptions {
+  participant?: Participant;
+}
+
+export function useIsMuted(source: Track.Source, options: UseIsMutedOptions = {}) {
+  const p = useEnsureParticipant(options.participant);
+  const [isMuted, setIsMuted] = React.useState(!!p.getTrack(source)?.isMuted);
+
+  React.useEffect(() => {
+    const listener = mutedObserver(p, source).subscribe(setIsMuted);
+    return () => listener.unsubscribe();
+  }, [p, source]);
+
+  return isMuted;
+}
