+++ packages/react/src/hooks/useLocalParticipant.ts
@@ -1,0 +1,48 @@
+import { ParticipantMedia, observeParticipantMedia } from '@livekit/components-core';
+import { TrackPublication, LocalParticipant } from 'livekit-client';
+import * as React from 'react';
+import { useRoomContext } from '../context';
+
+/**
+ * The useLocalParticipant hook the state of the local participant.
+ */
+export const useLocalParticipant = () => {
+  const room = useRoomContext();
+  const [localParticipant, setLocalParticipant] = React.useState(room.localParticipant);
+  const [isMicrophoneEnabled, setIsMicrophoneEnabled] = React.useState(
+    localParticipant.isMicrophoneEnabled,
+  );
+  const [isCameraEnabled, setIsCameraEnabled] = React.useState(
+    localParticipant.isMicrophoneEnabled,
+  );
+  const [isScreenShareEnabled, setIsScreenShareEnabled] = React.useState(
+    localParticipant.isMicrophoneEnabled,
+  );
+  const [microphoneTrack, setMicrophoneTrack] = React.useState<TrackPublication | undefined>(
+    undefined,
+  );
+  const [cameraTrack, setCameraTrack] = React.useState<TrackPublication | undefined>(undefined);
+
+  const handleUpdate = (media: ParticipantMedia<LocalParticipant>) => {
+    setIsCameraEnabled(media.isCameraEnabled);
+    setIsMicrophoneEnabled(media.isMicrophoneEnabled);
+    setIsScreenShareEnabled(media.isScreenShareEnabled);
+    setCameraTrack(media.cameraTrack);
+    setMicrophoneTrack(media.microphoneTrack);
+    setLocalParticipant(media.participant);
+  };
+  React.useEffect(() => {
+    const listener = observeParticipantMedia(localParticipant).subscribe(handleUpdate);
+    // TODO also listen to permission and metadata etc. events
+    return () => listener.unsubscribe();
+  }, [localParticipant]);
+
+  return {
+    isMicrophoneEnabled,
+    isScreenShareEnabled,
+    isCameraEnabled,
+    microphoneTrack,
+    cameraTrack,
+    localParticipant,
+  };
+};
