+++ packages/react/src/hooks/useParticipantPermissions.ts
@@ -1,0 +1,16 @@
+import * as React from 'react';
+import { Participant } from 'livekit-client';
+import { participantPermissionObserver } from '@livekit/components-core';
+import { useEnsureParticipant } from '../context';
+import { useObservableState } from '../helper/useObservableState';
+
+export interface UseParticipantPermissionsOptions {
+  participant?: Participant;
+}
+
+export function useParticipantPermissions(options: UseParticipantPermissionsOptions = {}) {
+  const p = useEnsureParticipant(options.participant);
+  const permissionObserver = React.useMemo(() => participantPermissionObserver(p), [p]);
+  const permissions = useObservableState(permissionObserver, p.permissions);
+  return permissions;
+}
