+++ packages/react/src/hooks/useParticipants.ts
@@ -1,0 +1,22 @@
+import { RoomEvent } from 'livekit-client';
+import { useLocalParticipant } from './useLocalParticipant';
+import { useRemoteParticipants } from './useRemoteParticipants';
+
+export interface UseParticipantsOptions {
+  /**
+   * To optimize performance, you can use the `updateOnlyOn` property to decide on what RoomEvents the hook updates.
+   * By default it updates on all relevant RoomEvents to keep the returned participants array up to date.
+   * The minimal set of non-overwriteable `RoomEvents` is: `[RoomEvent.ParticipantConnected, RoomEvent.ParticipantDisconnected, RoomEvent.ConnectionStateChanged]`
+   */
+  updateOnlyOn?: RoomEvent[];
+}
+
+/**
+ * The useParticipants hook returns all participants (local and remote) of the current room.
+ */
+export const useParticipants = (options: UseParticipantsOptions = {}) => {
+  const remoteParticipants = useRemoteParticipants(options);
+  const { localParticipant } = useLocalParticipant();
+
+  return [localParticipant, ...remoteParticipants];
+};
