+++ packages/react/src/hooks/useSortedParticipants.ts
@@ -1,0 +1,19 @@
+import { sortParticipants } from '@livekit/components-core';
+import { Participant } from 'livekit-client';
+import * as React from 'react';
+import { useSpeakingParticipants } from './useSpeakingParticipants';
+
+/**
+ * The useSortedParticipants hook returns the only the active speakers of all participants.
+ */
+export function useSortedParticipants(participants: Array<Participant>) {
+  const [sortedParticipants, setSortedParticipants] = React.useState(
+    sortParticipants(participants),
+  );
+  const activeSpeakers = useSpeakingParticipants();
+
+  React.useEffect(() => {
+    setSortedParticipants(sortParticipants(participants));
+  }, [activeSpeakers, participants]);
+  return sortedParticipants;
+}
