+++ packages/react/src/hooks/useSpeakingParticipants.ts
@@ -1,0 +1,14 @@
+import { activeSpeakerObserver } from '@livekit/components-core';
+import * as React from 'react';
+import { useRoomContext } from '../context';
+import { useObservableState } from '../helper';
+
+/**
+ * The useSpeakingParticipants hook returns the only the active speakers of all participants.
+ */
+export const useSpeakingParticipants = () => {
+  const room = useRoomContext();
+  const speakerObserver = React.useMemo(() => activeSpeakerObserver(room), [room]);
+  const activeSpeakers = useObservableState(speakerObserver, room.activeSpeakers);
+  return activeSpeakers;
+};
